Do some changes to the ticket model

+++ src/main/java/com/example/cinemaapp/Repository/RoomRepo.java
@@ -3,4 +3,4 @@ package com.example.cinemaapp.Repository;
 import com.example.cinemaapp.Model.Room;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoomRepo extends JpaRepository<Room, Long> {}
+public interface RoomRepo extends JpaRepository<Room, Integer> {}
